+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- the built-in ruler tool now shows the walk speed AP usage for the distance
+  measured
+
 ## [0.5.0] - 2021-10-03
 
 ### Added